release 1.0.59

+++ mobile/src/main/java/com/fnc/order/android/datacontroller/DcStaffs.java
@@ -56,7 +56,7 @@ public class DcStaffs extends DBHelper {
         db.close();
     }
 
-    public LinkedList<aStaffs> checkIfActive(Integer intEmpId) {
+    public LinkedList<aStaffs> checkIfActive(Long intEmpId) {
         SQLiteDatabase db = getReadableDatabase();
         String strQry = "SELECT * FROM " + Table.STAFFS_INACTIVE.getName() +
             " WHERE " + aStaffsKey.EMPID.getKey() + " = ?";
@@ -122,13 +122,13 @@ public class DcStaffs extends DBHelper {
 
     private aStaffs setStaffs(Cursor c) {
         aStaffs sl = new aStaffs(
-            c.getInt(c.getColumnIndex(aStaffsKey.EMPID.getKey())),
+            c.getLong(c.getColumnIndex(aStaffsKey.EMPID.getKey())),
             c.getString(c.getColumnIndex(aStaffsKey.REFEMPNO.getKey())),
             c.getString(c.getColumnIndex(aStaffsKey.EMPNO.getKey())),
             c.getString(c.getColumnIndex(aStaffsKey.EMAIL.getKey())),
             c.getString(c.getColumnIndex(aStaffsKey.NAME.getKey())),
-            c.getInt(c.getColumnIndex(aStaffsKey.BRANCH.getKey())),
-            c.getInt(c.getColumnIndex(aStaffsKey.JOBTITLE.getKey())),
+            c.getLong(c.getColumnIndex(aStaffsKey.BRANCH.getKey())),
+            c.getLong(c.getColumnIndex(aStaffsKey.JOBTITLE.getKey())),
             c.getString(c.getColumnIndex(aStaffsKey.PASS.getKey())),
             c.getString(c.getColumnIndex(aStaffsKey.ACTIVE.getKey())),
             c.getString(c.getColumnIndex(aStaffsKey.ISMOBILEADMIN.getKey()))
