+++ mobile/src/main/java/com/fnc/order/android/fragment/CustomerFragment.java
@@ -676,16 +676,16 @@ public class CustomerFragment extends Fragment implements VolleyCallback{
                             for (int i = 0; i < sArr.length(); i++) {
                                 JSONObject rowObj = sArr.getJSONObject(i);
                                 aStaffs sl = new aStaffs(
-                                        rowObj.getInt("empId"),
-                                        rowObj.getString("refempno").equals("null") ? "-1" : rowObj.getString("refempno"),
-                                        rowObj.getString("empNo"),
-                                        rowObj.getString("Email"),
-                                        rowObj.getString("name"),
-                                        rowObj.getInt("Branch"),
-                                        rowObj.getInt("Jobtitle"),
-                                        rowObj.getString("pass"),
-                                        rowObj.getString("active"),
-                                        rowObj.getString("ismobileadmin")
+                                    rowObj.getLong("empId"),
+                                    rowObj.getString("refempno").equals("null") ? "-1" : rowObj.getString("refempno"),
+                                    rowObj.getString("empNo"),
+                                    rowObj.getString("Email"),
+                                    rowObj.getString("name"),
+                                    rowObj.getLong("Branch"),
+                                    rowObj.getLong("Jobtitle"),
+                                    rowObj.getString("pass"),
+                                    rowObj.getString("active"),
+                                    rowObj.getString("ismobileadmin")
                                 );
                                 DcStaffs.getInstance(ctx).insertStaffs(sl);
                             }
@@ -827,33 +827,33 @@ public class CustomerFragment extends Fragment implements VolleyCallback{
                     for (int i = 0; i < objArr.length(); i++) {
                         JSONObject rowObj = objArr.getJSONObject(i);
                         aItemlist ail = new aItemlist(
-                                String.valueOf(rowObj.getString(aItemlistKey.INTEGRATION_RECID.getKey())),
-                                rowObj.getInt(aItemlistKey.RECID.getKey()),
-                                String.valueOf(rowObj.getString(aItemlistKey.OLD_SKU.getKey())),
-                                rowObj.getInt(aItemlistKey.BASEUNIT_RECID.getKey()),
-                                rowObj.getDouble(aItemlistKey.BASEUNIT_QTY.getKey()),
-                                rowObj.getString(aItemlistKey.ITEMNO.getKey()),
-                                rowObj.getString(aItemlistKey.ITEMNAME.getKey()),
-                                rowObj.getString(aItemlistKey.ITEMNAME_WUNIT.getKey()),
-                                rowObj.getDouble(aItemlistKey.QUANTITY_INUNIT.getKey()),
-                                rowObj.getString(aItemlistKey.DEPT.getKey()),
-                                rowObj.getString(aItemlistKey.UNIT.getKey()),
-                                rowObj.getInt(aItemlistKey.TBLUNIT_RECID.getKey()),
-                                rowObj.getInt(aItemlistKey.UNIT_TOCONVERT.getKey()),
-                                rowObj.getString(aItemlistKey.BARCODENO.getKey()),
-                                rowObj.getBoolean(aItemlistKey.F_BASE.getKey()) == true ? 1 : 0,
-                                rowObj.getString(aItemlistKey.D_ITEMDEPARTMENT_CODE.getKey()),
-                                String.valueOf(rowObj.getString(aItemlistKey.SELLING_PRICE.getKey())).equals("null")
-                                        ? 0.00 : rowObj.getDouble(aItemlistKey.SELLING_PRICE.getKey()),
-                                String.valueOf(rowObj.getString(aItemlistKey.COST_PRICE.getKey())).equals("null")
-                                        ? 0.00 : rowObj.getDouble(aItemlistKey.COST_PRICE.getKey()),
-                                rowObj.getString(aItemlistKey.TAXCODE.getKey()),
-                                String.valueOf(rowObj.getString(aItemlistKey.EXPENSE_ACCT.getKey())).equals("null")
-                                        ? 0 : rowObj.getInt(aItemlistKey.EXPENSE_ACCT.getKey()),
-                                String.valueOf(rowObj.getString(aItemlistKey.INCOME_ACCT.getKey())).equals("null")
-                                        ? 0 : rowObj.getInt(aItemlistKey.INCOME_ACCT.getKey()),
-                                rowObj.getString(aItemlistKey.DATA_VISIBILITY.getKey()),
-                                rowObj.getString(aItemlistKey.BARCODENO1.getKey())
+                            String.valueOf(rowObj.getString(aItemlistKey.INTEGRATION_RECID.getKey())),
+                            rowObj.getLong(aItemlistKey.RECID.getKey()),
+                            String.valueOf(rowObj.getString(aItemlistKey.OLD_SKU.getKey())),
+                            rowObj.getInt(aItemlistKey.BASEUNIT_RECID.getKey()),
+                            rowObj.getDouble(aItemlistKey.BASEUNIT_QTY.getKey()),
+                            rowObj.getString(aItemlistKey.ITEMNO.getKey()),
+                            rowObj.getString(aItemlistKey.ITEMNAME.getKey()),
+                            rowObj.getString(aItemlistKey.ITEMNAME_WUNIT.getKey()),
+                            rowObj.getDouble(aItemlistKey.QUANTITY_INUNIT.getKey()),
+                            rowObj.getString(aItemlistKey.DEPT.getKey()),
+                            rowObj.getString(aItemlistKey.UNIT.getKey()),
+                            rowObj.getInt(aItemlistKey.TBLUNIT_RECID.getKey()),
+                            rowObj.getInt(aItemlistKey.UNIT_TOCONVERT.getKey()),
+                            rowObj.getString(aItemlistKey.BARCODENO.getKey()),
+                            rowObj.getBoolean(aItemlistKey.F_BASE.getKey()) == true ? 1 : 0,
+                            rowObj.getString(aItemlistKey.D_ITEMDEPARTMENT_CODE.getKey()),
+                            String.valueOf(rowObj.getString(aItemlistKey.SELLING_PRICE.getKey())).equals("null")
+                                    ? 0.00 : rowObj.getDouble(aItemlistKey.SELLING_PRICE.getKey()),
+                            String.valueOf(rowObj.getString(aItemlistKey.COST_PRICE.getKey())).equals("null")
+                                    ? 0.00 : rowObj.getDouble(aItemlistKey.COST_PRICE.getKey()),
+                            rowObj.getString(aItemlistKey.TAXCODE.getKey()),
+                            String.valueOf(rowObj.getString(aItemlistKey.EXPENSE_ACCT.getKey())).equals("null")
+                                    ? 0 : rowObj.getInt(aItemlistKey.EXPENSE_ACCT.getKey()),
+                            String.valueOf(rowObj.getString(aItemlistKey.INCOME_ACCT.getKey())).equals("null")
+                                    ? 0 : rowObj.getInt(aItemlistKey.INCOME_ACCT.getKey()),
+                            rowObj.getString(aItemlistKey.DATA_VISIBILITY.getKey()),
+                            rowObj.getString(aItemlistKey.BARCODENO1.getKey())
                         );
                         DcAitemlist.getInstance(ctx).insertaItemlist(ail);
                     }
@@ -934,16 +934,16 @@ public class CustomerFragment extends Fragment implements VolleyCallback{
                                     for (int i = 0; i < sArr.length(); i++) {
                                         JSONObject rowObj = sArr.getJSONObject(i);
                                         aStaffs sl = new aStaffs(
-                                                rowObj.getInt("empId"),
-                                                rowObj.getString("refempno").equals("null") ? "-1" : rowObj.getString("refempno"),
-                                                rowObj.getString("empNo"),
-                                                rowObj.getString("Email"),
-                                                rowObj.getString("name"),
-                                                rowObj.getInt("Branch"),
-                                                rowObj.getInt("Jobtitle"),
-                                                rowObj.getString("pass"),
-                                                rowObj.getString("active"),
-                                                rowObj.getString("ismobileadmin")
+                                            rowObj.getLong("empId"),
+                                            rowObj.getString("refempno").equals("null") ? "-1" : rowObj.getString("refempno"),
+                                            rowObj.getString("empNo"),
+                                            rowObj.getString("Email"),
+                                            rowObj.getString("name"),
+                                            rowObj.getLong("Branch"),
+                                            rowObj.getLong("Jobtitle"),
+                                            rowObj.getString("pass"),
+                                            rowObj.getString("active"),
+                                            rowObj.getString("ismobileadmin")
                                         );
                                         DcStaffs.getInstance(ctx).insertStaffs(sl);
                                     }
